베이스 코드로 변경

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -1,71 +1,65 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.binarycontent.BinaryContentDTO;
-import com.sprint.mission.discodeit.dto.binarycontent.CreateBinaryContentRequest;
+import com.sprint.mission.discodeit.dto.data.BinaryContentDto;
+import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
-import com.sprint.mission.discodeit.exception.ErrorCode;
-import com.sprint.mission.discodeit.exception.ServiceException;
+import com.sprint.mission.discodeit.mapper.BinaryContentMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.service.BinaryContentService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.UUID;
-
-@Service
 @RequiredArgsConstructor
+@Service
 public class BasicBinaryContentService implements BinaryContentService {
 
   private final BinaryContentRepository binaryContentRepository;
-  private final ModelMapper modelMapper;
+  private final BinaryContentMapper binaryContentMapper;
   private final BinaryContentStorage binaryContentStorage;
 
   @Transactional
   @Override
-  public BinaryContentDTO create(CreateBinaryContentRequest request) throws IOException {
-    String filename = request.fileName();
+  public BinaryContentDto create(BinaryContentCreateRequest request) {
+    String fileName = request.fileName();
     byte[] bytes = request.bytes();
     String contentType = request.contentType();
-    BinaryContent binaryContent = BinaryContent.builder()
-            .fileName(filename)
-            .contentType(contentType)
-            .size((long) bytes.length)
-            .build();
+    BinaryContent binaryContent = new BinaryContent(
+        fileName,
+        (long) bytes.length,
+        contentType
+    );
+    binaryContentRepository.save(binaryContent);
+    binaryContentStorage.put(binaryContent.getId(), bytes);
 
-    BinaryContent saveBinaryContent = binaryContentRepository.save(binaryContent);
-
-    binaryContentStorage.put(saveBinaryContent.getId(), bytes);
-
-    return modelMapper.map(binaryContent, BinaryContentDTO.class);
+    return binaryContentMapper.toDto(binaryContent);
   }
 
-  @Transactional(readOnly = true)
   @Override
-  public BinaryContentDTO find(UUID binaryContentId) {
-    BinaryContent binaryContent = binaryContentRepository.findById(binaryContentId)
-            .orElseThrow(() -> new ServiceException(ErrorCode.CANNOT_FOUND_PROFILE));
-
-    return modelMapper.map(binaryContent, BinaryContentDTO.class);
+  public BinaryContentDto find(UUID binaryContentId) {
+    return binaryContentRepository.findById(binaryContentId)
+        .map(binaryContentMapper::toDto)
+        .orElseThrow(() -> new NoSuchElementException(
+            "BinaryContent with id " + binaryContentId + " not found"));
   }
 
-  @Transactional(readOnly = true)
   @Override
-  public List<BinaryContentDTO> findAllByIdIn(List<UUID> binaryContentIds) {
-    List<BinaryContent> binaryContents = binaryContentRepository.findAllIdIn(binaryContentIds);
-
-    return binaryContents.stream()
-            .map(binaryContent -> modelMapper.map(binaryContent, BinaryContentDTO.class))
-            .toList();
+  public List<BinaryContentDto> findAllByIdIn(List<UUID> binaryContentIds) {
+    return binaryContentRepository.findAllById(binaryContentIds).stream()
+        .map(binaryContentMapper::toDto)
+        .toList();
   }
 
   @Transactional
   @Override
   public void delete(UUID binaryContentId) {
+    if (!binaryContentRepository.existsById(binaryContentId)) {
+      throw new NoSuchElementException("BinaryContent with id " + binaryContentId + " not found");
+    }
     binaryContentRepository.deleteById(binaryContentId);
   }
 }
